Add tracing support
Instead of println'ing everything, this commit switches the code base to
using the tracing library. There are now proper log levels and multiple
logging output formats. A bunch of new debug and trace-level messages
have also been added to help with future troubleshooting.

The output is a bit uglier than before due to liberal use of spans and
structured logging. We can tweak the message formatting later if needed.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>